Support opam findlib + system compiler
When findlib is installed via opam as a normal user, it cannot
install 'topfind' in the system compiler's directory. Instead
opam will set the environment variable OCAML_TOPLEVEL_PATH to point
to it and add some runes to .ocamlinit.

This patch adds the same runes to the generated OCaml code before the
'#use topfind' line.

See ocaml/opam#45